+++ ElvUI_LuckyoneUI/Core/Config.lua
@@ -70,10 +70,11 @@ function L1UI:Config()
 	-- Installer & Update
 	L1UI.Options.args.setup = ACH:Group('', nil, 1, nil, function(info) return E.db.L1UI.general[info[#info]] end, function(info, value) E.db.L1UI.general[info[#info]] = value end)
 	L1UI.Options.args.setup.inline = true
-	L1UI.Options.args.setup.args.auto_update = ACH:Toggle(L["Auto Update"], L["Automatically update after a new release."], 1)
-	L1UI.Options.args.setup.args.installer = ACH:Execute(L["Install"], L["Re-Run the installation process."], 2, function() PI:Queue(L1UI.InstallerData) E:ToggleOptions() end)
-	L1UI.Options.args.setup.args.updateMain = ACH:Execute(L["Update Main Layout"], L["Update Main layout to LuckyoneUI version: "]..Private.Version, 3, function() L1UI:UpdateLayout('main') end, nil, true)
-	L1UI.Options.args.setup.args.updateHealing = ACH:Execute(L["Update Healing Layout"], L["Update Healing layout to LuckyoneUI version: "]..Private.Version, 4, function() L1UI:UpdateLayout('healing') end, nil, true)
+	L1UI.Options.args.setup.args.installer = ACH:Execute(L["Install"], L["Re-Run the installation process."], 1, function() PI:Queue(L1UI.InstallerData) E:ToggleOptions() end)
+	L1UI.Options.args.setup.args.updateMain = ACH:Execute(L["Update Main Layout"], L["Update Main layout to LuckyoneUI version: "]..Private.Version, 2, function() L1UI:UpdateLayout('main') end, nil, true)
+	L1UI.Options.args.setup.args.updateHealing = ACH:Execute(L["Update Healing Layout"], L["Update Healing layout to LuckyoneUI version: "]..Private.Version, 3, function() L1UI:UpdateLayout('healing') end, nil, true)
+	L1UI.Options.args.setup.args.updateResolution = ACH:Execute(L["1080p > 1440p"], L["Scale the default 1080p to 1440p"]..'|cffC80000\n\nTHIS IS ONLY FOR MAIN AND SUPPORT LAYOUTS (NOT HEALING)|r', 4, function() L1UI:UpdateResolution() end, nil, true)
+	L1UI.Options.args.setup.args.auto_update = ACH:Toggle(L["Auto Update"], L["Automatically update after a new release."], 5)
 
 	-- Spacer
 	L1UI.Options.args.header = ACH:Spacer(2, 'full')