+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### Version 2.57 [In Development]
+### Version 2.57 [01-05-2024]
 **All WoW Versions:**  
 Added option to scale Main layout and Support layout from 1080p to 1440p  
 Future layouts starting in patch 11.0 will be 1440p by default